HTTPS for common source code distribution domains
All links to the following are how https
* pypy.python.org
* rubygems.org
* launchpad.net
* github.com
* bitbucket.org
* freedesktop.org
  * Excluding xorg.freedesktop.org
  * Excluding tango.freedesktop.org

All of these have been renamed to their canoncial form:
* www.github.com -> github.com
* www.bitbucket.org -> bitbucket.org
* pkg-config.freedesktop.org -> pkgconfig.freedesktop.org
* standards.freedesktop.org -> specifications.freedesktop.org

During the automated testing of all modified packages the following SRC_URIs
were discovered to have changed. Each one is documented under the action taken.
Many of these ebuilds were not otherwise modified by this pull request and
simply happened to live next to a newer version that was modified.

Fixed ebuilds:
* app-admin/elektra-0.7.1
* app-admin/packagekit-base-0.8.15-r1
* app-text/libabw-libabw-0.0.2
* app-text/libetonyek-0.0.3
* app-text/libmspfub-0.0.6
* media-libs/libcdr-0.0.14
* media-libs/libcdr-0.0.16
* media-libs/libfreehand-0.0.0
* media-libs/libvisio-0.0.30
* x11-misc/xrestop-0.4

Ignored ebuilds because the upstream tarball has a new hash:
* app-misc/evtest-1.29
* app-misc/evtest-1.30
* dev-embedded/scratchbox2-2.0-r1
* dev-ml/cairo-ocaml-1.2.0
* dev-python/rospkg-1.0.38qq
* dev-python/vcstools-0.1.38
* dev-util/rosdep-0.11.4
* dev-util/wstool-0.1.12
* net-libs/libqmi-1.0.0
* sys-auth/libfprint-0.4.0
* sys-auth/libfprint-0.5.0
* sys-auth/libfprint-0.5.1

Ignored ebuilds because the upstream tarball has been deleted:
* app-admin/packagekit-qt4-0.8.8
* dev-libs/liblazy-0.2
* dev-python/pyx-0.12.1-r1
* dev-python/pyx-0.13
* dev-python/pyx-0.14
* dev-util/pkgconfig-openbsd-20130507-r1
* sys-auth/consolekit-0.4.6
* x11-libs/xvba-video-0.8.0-r3
* x11-misc/driconf-0.9.1-r1
* x11-misc/xdg-utils-1.1.0_rc2-r1

All of the problems that we ignored were deemed to not be a regression and
people using those ebuilds should not be adversely affected, since the expected
tarballs are available on the mirrors.